\name{decoReport}
\alias{decoReport}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Report of DECO analyisis
}
\description{
This report includes a PDF file with several different representations from the analysis. Additionally, a text file (tabulated-delimited) including \emph{featureTable} slot from a \bold{deco} R object and a text file with \emph{\bold{h}} statistics per feature and subclass found were both generated. They will be created inside current working directory.
}
\usage{
decoReport(deco, sub, id = NA, pdf.file = "decoReport.pdf",
                 info.sample = NA, info.feature = NA,
                 cex.samples = 1.2, cex.legend = 0.9,
                 cex.names = 0.8, print.annot = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{deco}{
a "deco" R object generated by 'decoNSCA' R function.
}
  \item{sub}{
a R object generated by 'decoRDA' R function.
}
  \item{id}{
character vector indicating what original IDs from input data should be highlighted and plot in 'Profile' section. If all IDs found have to be plot, input all corresponding IDs here.
}
  \item{pdf.file}{
path or name of pdf where report should be generated.
}
  \item{info.sample}{
a factor or data.frame with relevant information of samples to be plot on 'Heatmap'.
}
  \item{info.feature}{
a factor or data.frame with relevant information of genes or features to be plot on 'Heatmap'.
}
  \item{cex.samples}{
numerical value giving the amount of magnification of sample related information, i.e. sample points in 'Profile' plots.
}
  \item{cex.legend}{
numerical value giving the amount of magnification of legend.
}
  \item{cex.names}{
numerical value giving the amount of magnification to gene names or sample names.
}
  \item{print.annot}{
logical indicating if annotation (gene SYMBOL) should be printed instead original IDs. This option needs previous annotation within 'decoRDA' function or late annotation binding information to @featureTable slot of 'deco' R object.
}
}
\value{
Returns a 'pdf' file including relevant information about all DECO algorithm analysis.
}
\author{
Francisco Jose Campos Laborie. <fjcamlab@usal.es>
}
\seealso{
Vignette of R package explains all different plots included within the pdf.

\code{\link[deco]{decoRDA}}, \code{\link[deco]{decoNSCA}}
}
\examples{
# Phenotypical data from TCGA RNAseq samples.
data(ALCLdata)
head(colData(ALCL))

#####################################################
# PDF report with gene-sample patterns or subgroups #
#####################################################
## Generate PDF report with relevant information and several plots.
# It will also generate a txt (tab-delimitated) table including featureTable.
# Both objects will be saved in your working directory if any path is given in PDF name.

## Binary example (ALK+ vs ALK-)
# decoReport(deco.results.ma, sub.ma.3r.1K,
#            pdf.file = "report_example_microarray_binary.pdf",
#            info.sample = as.data.frame(colData(ALCL))[,8:10],
#            cex.names = 0.3, print.annot = TRUE)

## Unsupervised example (no classes)
# decoReport(deco.results.ma.uns, sub.ma.3r.1K.uns,
#            pdf.file = "report_example_microarray_unsupervised.pdf",
#            info.sample = as.data.frame(colData(ALCL))[,8:10],
#            cex.names = 0.3, print.annot = TRUE)

## Multiclass example (ALK+ vs PTCL vs ALK-(noPTCL))
# decoReport(deco.results.ma.multi, sub.ma.3r.1K.multi,
#            pdf.file = "report_example_microarray_multi.pdf",
#            info.sample = as.data.frame(colData(ALCL))[,8:10],
#            cex.names = 0.3, print.annot = TRUE)


}
