\name{plotAssociationH}
\alias{plotAssociationH}
\title{
Plot to visualize association among DECO subclasses and sample information (phenotype).
}
\description{
This function returns three plots showing the association among different sample subclasses found by DECO and any sample information of interest. This information can be the initial classes or any new included sample information. 
}
\usage{
plotAssociationH(deco, info.sample)
}
\arguments{
  \item{deco}{
a "deco" R object generated by 'decoNSCA' R function.
}
  \item{info.sample}{
a character vector or factor including the sample information.
}
}
\details{
This function provides a simple comparison of any phenotype characteristic and new DECO subclasses.
}
\value{
Violin plot: distribution of average h-statistic values (per feature) within the newly provided categories of samples, splitting by the subsets of features associated to each DECO subclass. 

Left-heatmap: frequency table of \emph{samples} among DECO subclasses and newly provided sample categories.
Right-heatmap: frequency table of \emph{features} among DECO subclasses and newly provided sample categories.
}
\author{
Francisco Jose Campos Laborie. <fjcamlab@gmail.com>
}
\seealso{
\code{\link[deco]{decoRDA}}, \code{\link[deco]{decoNSCA}}
}
\examples{
#### Further information in DECO vignette(s).
## Run after decoRDA and decoNSCA steps on 'ALCLdata'
data(ALCLdata)
ALCL

### Sample source information
info.sample <- as.data.frame(colData(ALCL))[,"Sample.origin"]
names(info.sample) <- rownames(colData(ALCL))

#### Comparing DECO subclasses against source of samples.
plotAssociationH(deco.results.ma, info.sample)
}
