\name{plotGainingH}
\alias{plotGainingH}
\title{
Gaining plots for using h-statistic instead original omic data.
}
\description{
This function returns three plots for a particular feature(s), showing the correlation between the h-statistic calculated by DECO and the original omic data. 
}
\usage{
plotGainingH(deco, data, ids, print.annot = FALSE, 
             orig.classes = TRUE)
}
\arguments{
  \item{deco}{
a "deco" R object generated by 'decoNSCA' R function.
}
  \item{data}{
a matrix of normalized omic data with 'f' features (rows) by 'n' samples (columns).
}
  \item{ids}{
character vector indicating what original IDs from input data should be highlighted and plot in 'Profile' section. If all IDs found have to be plot, input all corresponding IDs here.
}
  \item{print.annot}{
logical indicating if annotation (gene SYMBOL) should be printed instead original IDs. This option needs previous annotation within 'decoRDA' function or late annotation binding information to @featureTable slot of 'deco' R object.
}  \item{orig.classes}{
logical indicating if original sample categories compared (orig.classes = TRUE) or DECO subclasses (orig.classes = FALSE) will be represented.
}
}
\details{
Diagnostic plot associating omic profile and h-statistic profile of any significant feature found by DECO.
}
\value{
Three plots are returned:
Boxplot: distribution of omic data and h-statistic values per DECO subclass. 
Top-left: parametric correlation between omic data and h-statistic per sample.
Top-right: non-parametric correlation (ranking) between omic data and h-statistic per sample.
}
\author{
Francisco Jose Campos Laborie. <fjcamlab@gmail.com>
}
\seealso{
\code{\link[deco]{decoRDA}}, \code{\link[deco]{decoNSCA}}
}
\examples{
#### Further information in DECO vignette(s).
## Run after decoRDA and decoNSCA steps on 'ALCLdata'
data(ALCLdata)
ALCL

## Feature to represent
id <- featureTable(deco.results.ma)[1, "ID"]

#### Comparing DECO subclasses against source of samples.
plotGainingH(deco.results.ma, data = assay(ALCL), ids = id, 
             print.annot = FALSE, orig.classes = FALSE)
}

