\name{plotDECOProfile}
\alias{plotDECOProfile}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Feature profile from DECO analysis
}
\description{
Plot feature profile(s) in a separate PDF from decoReport R function.
}
\usage{
plotDECOProfile(deco, id, data, pdf.file = NA, plot.h = FALSE,
                      info.sample = NA, print.annot = TRUE,
                      cex.legend = 1.1, cex.names = 1, cex.samples = 1)
}
\arguments{
  \item{deco}{
a "deco" R object generated by 'decoNSCA' R function.
}
  \item{id}{
character vector indicating what original IDs from input data should be highlighted and plot in 'Profile' section. If all IDs found have to be plot, input all corresponding IDs here.
}
  \item{data}{
input matrix of normalized data with 'f' features (rows) by 's' samples (columns).
}
  \item{pdf.file}{
path or name of pdf where report should be generated.
}
  \item{plot.h}{
logical indicating if points corresponding to 'h' relative values should be plotted.
}
  \item{info.sample}{
a factor or data.frame with relevant information of samples to be plot on 'Heatmap'.
}
  \item{print.annot}{
logical indicating if annotation (gene SYMBOL) should be printed instead original IDs. This option needs previous annotation within 'decoRDA' function or late annotation binding information to @featureTable slot of 'deco' R object.
}
  \item{cex.legend}{
numerical value giving the amount of magnification of legend.
}
  \item{cex.names}{
numerical value giving the amount of magnification to gene names or sample names.
}
  \item{cex.samples}{
numerical value giving the amount of magnification of sample related information, i.e. sample points in 'Profile' plots.
}
}
\value{
Returns a new PDF.
}
\author{
Francisco Jose Campos Laborie. <fjcamlab@usal.es>
}
\seealso{
\code{\link[deco]{decoNSCA}}, \code{\link[limma]{voom}}
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
